Add simple caching to merge_styles
Profiling show the majority of the work is going on in the
'merge_styles' function and more explicitly in the CSS parsing so we
cache the result of the parsing in a dictionary against the CSS string
to speed that up.

Once that was cached, the next main bottle neck was the call to:

    parsed.getVariableValue(key)

So we cache that as well.

The caches are currently module level. This is not recommended but
briefly introduced for testing purposes.