Place variable cache within styles cache
We were using a combined key for the variable_value_cache but that is to
get around the fact that the variable_value_cache should be associated
with that CSS text so we store the variable_value_cache in the
styles_cache against the 'csstext' key so now our variable_value_cache
lookups occur against individual small dictionaries instead of a large
global dictionary.

I assume this is faster but honestly I doesn't make a big different so
really it just feels neater.