Use local non-module level cache
We switch from a global module level cache to declaring the cache in the
'transform' method and passing it into the calls to 'merge_styles'.

Global variables are bad for all kinds of reasons so hopefully this
doesn't need too much explanation.

We update the tests as well to handle the new argument.