Fix tests for travis
We switch to assertTrue as there is no form of assertLess available on
Python 2.6.

We test if we're on PyPy and assert a much larger test time as it seems
to take a long time. I've done no investigations into this. I don't know
if my changes have caused this but I have no reason to believe this.

We throw an exception for unrecognised platforms so that we don't
silently pass the test or incorrectly fail it.